[mlir][tensor] Add TilingInterface implementation for expand_shape
Because of the value semantic nature of the tensor type,
`tensor.expand_shape` can be thought of as a copy from a tensor of lower
rank to one of higher rank. This implements the tiling interface for
this operation using the shape of the result as the iteration bounds.

For now the tiling implementation only supports cases where the
expand_shape can be rewritten as a single expand shape. In particular,
the offsets and sizes into the iteration space must be convertible to
a single offset and size per reassociated group.